Add PWM controlled piezo methods
Driving the piezo this way allows it to be done entirely in the
background (no need for multitasking here, yet). Also, we're not
dropping into an interrupt routine to pulse the latch as Eric had done
previously in some of his example code.

tone_duration currently does nothing, it will need to be expanded by the
main interrupt driven loop (currently in dev/leds.c, but could possibly
be moved into a separate interrupts.c or the like). It will probably
have to be coupled, and will likely be in the screen redraw routine (for
short duration sounds) and not the scheduler (we don't really want cruft
in there anyway).