Filled in the serial code from eric's example
Also added a function to print strings. At the moment, if you use this, it looks like it doesn't hang up on the connection properly, when you go to flash it next it responds with an error, indicating that it's busy. So, I would only use it like it is for debugging