Optimise binaries for space
Added -Os flag to makeflash. It was needed for some example
multithreading code that was clobbering itself. 2K SRAM isn't much.
Also, it turns out that constant global arrays are in the SRAM, not the
flash (I'd expected them to be in .text, i.e. EEPROM). Harvard
architecture strikes again!